requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-hashid-field==3.1.3
django-mock-queries==2.1.5
django-rest-framework==0.1.0
djangorestframework==3.12.1
drf-yasg==1.17.1
factory-boy==3.0.1
Faker==4.1.4
hashids==1.3.1
idna==2.10
inflection==0.5.1
iniconfig==1.0.1
isort==5.5.3
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
model-bakery==1.0.2
openapi-codec==1.3.2
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.0
pytest-django==3.10.0
pytest-mock==3.3.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
